Send user_logged_in event when a token is requested
This way the "last_logged_in" column in the auth_user table
is updated properly and we can use that to identify "active"
users this year.
It turned out the is_active column is not suitable for that
because it's used to lock out users - they cannot log in
when that field is set to false.