Make sure test users are always considered "active"
At some places in the code, we filter on only "active" users,
which means not only those user which actually have the active
flag set to true, but also only to those users, who have logged
in this year.
This was not taken care of in the test utils method which creates
test users. This commit sets the last_login property to now()
and therefore makes sure that test users are never filtered out.