Upgrade prod requirements & scripts
Adjusted the required Python packages for production
to what the DjangoEurope installer sets when installing
a pristine Django 3 project.

Updated the corresponding remote scripts for starting
the app server etc. accordingly (by comparing once again
to the scripts that the installer has created).

This can only be properly tested on the demo/prod server
and additional commits to fix broken startup behaviour
are possible.