Adjust DEMO Nginx to temporarily redirect to PROD
This is due to the error we made in the welcome e-mail to accidentally
include links to the DEMO system. This confused some users and even led
to two users registering on the DEMO.

Hence, we temporarily redirect all traffic to the PROD to avoid further
mis-registrations. This is a temporary action.