Do not send author e-mail as "From"
In the contact form, so far the e-mail address from the author was set
as "From", as this seemed to be convenient. The idea was that we get the
message as if it was just send as regular e-mail from the author.

However, this does not seem to be allowed, as I got server errors
stating SMTPRecipientsRefused ... Sender address rejected: invalid
sender domain.

This commit changes the behaviour to instead just use our default
address as sender and include the author's name and e-mail in the e-mail
body.